Move values(), extend() and clone() to core/util/object
